Refactor create invoice form
+ extract invoice form & invoice confirmation to separate components
+ clean up code
+ minor improvements
